test: extract value control buttons tests
