+++ packages/number-field/test/value-control-buttons-lit.test.js
@@ -1,0 +1,2 @@
+import '../src/vaadin-lit-number-field.js';
+import './value-control-buttons.common.js';
