+++ packages/number-field/test/value-control-buttons-polymer.test.js
@@ -1,0 +1,2 @@
+import '../src/vaadin-number-field.js';
+import './value-control-buttons.common.js';
